+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### vNEXT
 
 * Remove dependency on `graphql-subscription` and use an interface for PubSub [PR #295](https://github.com/apollographql/graphql-tools/pull/295)
+* Support schema AST as a type definition input [PR #300](https://github.com/apollographql/graphql-tools/pull/300)
 
 ### v0.10.1
 * Update dependencies [PR #287](https://github.com/apollographql/graphql-tools/pull/287)